+++ Cargo.toml
@@ -4,6 +4,8 @@ version = "1.1.2"
 authors = ["Chris233 h-chris233@qq.com"]
 edition = "2021"
 readme = "README.md"
+repository = "https://github.com/H-Chris233/pingdown/"
+default-run = "pingdown"
 
 [dependencies]
 clap = { version = "4.5.30", features = ["derive"] }
